Create Challenge3
Write a function that attaches click listeners to all the buttons in the top menu. 

The click handler should first check whether or not a document cookie named ‘browserClick’ has been set on the page. If the cookie has not been set, the function should set the cookie to the text of the button clicked; if it has been set, the function should trigger an alert that displays the cookie value.